requirements.txt
dash
pandas
plotly.express
dash_bootstrap_components
nutritionix